requirements.txt
fastapi
matplotlib
numpy
opencv_python
pandas
Pillow
psutil
Requests
scipy
seaborn
torch
torchvision
ultralytics
ultralytics_thop
uvicorn[standard]==0.30.1
x2paddle
uvloop 
httpx

